Move all code about accessor / buffer / bufferview / sparse
