+++ addons/io_scene_gltf2/blender/imp/rig/armature.py
@@ -23,7 +23,7 @@
 
 import bpy
 from mathutils import Vector, Matrix, Quaternion
-from ..buffer import *
+from ....io.com.gltf2_io_accessor import *
 from ..util.conversion import * #TODO: move to com
 
 class Skin():
